+++ main.c
@@ -3,18 +3,17 @@
 #include <stdlib.h>
 #include "interface.h"
 #include "logic.h"
-#include "lista.h"
 
 #define BUF_SIZE 1024
 
 
 int main() {
-    srand(time(NULL));
+    srand((unsigned int) time(NULL));
     printf("\n      -----------------------\n\tWelcome to RASTROS!\n      -----------------------");
     ESTADO *e = inicializar_estado();
     printBoard(e);
     while (1) {
-        if ((int) interpretador(e) == 2) {
+        if (interpretador(e) == 2) {
             break;
         }
     }
